Move COM-stuff to a separate file
